# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My APIs",
        "contact": {
            "name": "SamSalvatico",
            "email": "samuele.salvatico@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/tree_node_names": {
            "get": {
                "tags": [
                    "TreeNodeNames"
                ],
                "summary": "Get all TreeNodeNames",
                "description": "Return all TreeNodeNames that matches request",
                "operationId": "getTreeNodeNames",
                "parameters": [
                    {
                        "name": "criteria",
                        "in": "query",
                        "description": "Some optional other parameter",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TreeNodeName"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                },
                "security": [
                    {
                        "oauth2Auth": [
                            "tree_node_names_view"
                        ]
                    }
                ]
            },
            "post": {
                "tags": [
                    "TreeNodeNames"
                ],
                "summary": "Add a new TreeNodeName",
                "description": "",
                "operationId": "addTreeNodeName",
                "requestBody": {
                    "description": "TreeNodeName object that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/xml": {
                            "schema": {
                                "$ref": "#/components/schemas/TreeNodeName"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TreeNodeName"
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid TreeNodeName"
                    },
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TreeNodeName"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "oauth2Auth": [
                            "tree_node_names_create"
                        ]
                    }
                ]
            }
        },
        "/api/v1/tree_node_names/{id}": {
            "get": {
                "tags": [
                    "TreeNodeNames"
                ],
                "summary": "Find TreeNodeName by ID",
                "description": "Returns a single TreeNodeName",
                "operationId": "getTreeNodeName",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of TreeNodeName to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TreeNodeName"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Item not found"
                    }
                },
                "security": [
                    {
                        "oauth2Auth": [
                            "tree_node_names_view"
                        ]
                    }
                ]
            },
            "put": {
                "tags": [
                    "TreeNodeNames"
                ],
                "summary": "Update an existing TreeNodeName",
                "description": "",
                "operationId": "updateTreeNodeName",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "TreeNodeName id to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "TreeNodeName object that needs to be updated into the store",
                    "required": true,
                    "content": {
                        "application/xml": {
                            "schema": {
                                "$ref": "#/components/schemas/TreeNodeName"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TreeNodeName"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TreeNodeName"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Item not found"
                    },
                    "405": {
                        "description": "Invalid TreeNodeName"
                    }
                },
                "security": [
                    {
                        "oauth2Auth": [
                            "tree_node_names_edit"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "TreeNodeNames"
                ],
                "summary": "Deletes an TreeNodeName",
                "description": "",
                "operationId": "deleteTreeNodeName",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "TreeNodeName id to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "TreeNodeName not found"
                    }
                },
                "security": [
                    {
                        "oauth2Auth": [
                            "tree_node_names_delete"
                        ]
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "TreeNodeName": {
                "required": [
                    "tree_node_id",
                    "language",
                    "node_name"
                ],
                "properties": {
                    "id": {
                        "description": "id",
                        "type": "integer",
                        "nullable": "false"
                    },
                    "tree_node_id": {
                        "description": "tree_node_id",
                        "type": "integer",
                        "nullable": "false"
                    },
                    "language": {
                        "description": "language",
                        "type": "string",
                        "nullable": "false"
                    },
                    "node_name": {
                        "description": "node_name",
                        "type": "string",
                        "nullable": "false"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "TreeNodeName"
                }
            }
        }
    }
}